feat: 입력 문자열 최대 길이 제한 추가
상용 공학용 계산기의 Input Rules 를 참고하여 입력 길이 제한 추가
- 참고 문서: [Casio Basic InputRules](https://bit.ly/4eNAR5H)

계산기 숫자 자료구조 변경
- int, long 최대 값을 넘는 숫자가 올 수 있기 때문에 BigInteger 사용

+++ src/main/java/calculator/numbercalculator/NumberCalculator.java
@@ -3,6 +3,7 @@ package calculator.numbercalculator;
 import calculator.numbercalculator.io.InputHandler;
 import calculator.numbercalculator.io.OutputHandler;
 import calculator.numbercalculator.number.Numbers;
+import java.math.BigInteger;
 
 public class NumberCalculator {
     private final InputHandler inputHandler;
@@ -18,7 +19,7 @@ public class NumberCalculator {
     public void run() {
         String userInput = inputHandler.getUserInput();
         Numbers numbers = numberExtractor.extractNumbers(userInput);
-        int result = numbers.calculateTotalValue();
+        BigInteger result = numbers.calculateTotalValue();
         outputHandler.showCalculateResult(result);
     }
 }
