+++ src/main/java/calculator/numbercalculator/io/OutputHandler.java
@@ -1,8 +1,10 @@
 package calculator.numbercalculator.io;
 
+import java.math.BigInteger;
+
 public class OutputHandler {
 
-    public void showCalculateResult(int result) {
+    public void showCalculateResult(BigInteger result) {
         System.out.printf("결과 : %d", result);
     }
 }
